storage: Add storage-configuration.yml mode . Implement READ_ONLY mode. #TASK-2495

+++ opencga-storage/opencga-storage-hadoop/opencga-storage-hadoop-core/src/test/java/org/opencb/opencga/storage/hadoop/variant/VariantHbaseTestUtils.java
@@ -319,6 +319,9 @@ public class VariantHbaseTestUtils {
     public static void printMetaTable(VariantHadoopDBAdaptor dbAdaptor, Path outDir) throws IOException {
 //        HBaseFileMetadataDBAdaptor fileDBAdaptor = dbAdaptor.getVariantFileMetadataDBAdaptor();
         String metaTableName = dbAdaptor.getTableNameGenerator().getMetaTableName();
+        if (!dbAdaptor.getHBaseManager().tableExists(metaTableName)) {
+            return;
+        }
         Path fileName = outDir.resolve("meta." + metaTableName + '.' + TimeUtils.getTimeMillis() + ".txt");
         try (
                 FileOutputStream fos = new FileOutputStream(fileName.toFile()); PrintStream out = new PrintStream(fos)
